+++ src/DigitalDriveShaft/sim/cylindrical/geometry.py
