+++ src/DigitalDriveShaft/sim/cylindrical/mesh.py
